refactor azure config into its own file
